Make nav selection a hook and fix blog selection for individual post
